+++ CHANGELOG.md
@@ -8,6 +8,7 @@ v0.8.5
 ---------------------
 * Issue #19: Updated README about how to access the site
 * Issue #20: Admin panel chrome console error: not working adminlte js
+* Issue #22: "Clickjacking" attack fix
 
 v0.8.4
 ---------------------